[CATACLYSM]
[Ovahlord] https://github.com/Ovahlord/TrinityCore
- DB: fixed naming of last db update
- DB/Movement: fixed hover movement for Unstable Twilight, Frostfire Orb and Flame Orb

[REPACK] - lasyan3's patches
- Custom.SpeedGame
Set it to a number upper than 1 to speed up realm timer

- Custom.NoCastTime
Set it to 1 to disable cast time

- Custom.NoCooldown
Set it to 1 to disable cooldown

- Custom.HurtInRealTime
Set it to 1 to disable autohurt with weapon

- Custom.AttackSpeedForPlayer
- Custom.AttackSpeedForMobs
Set it to a number upper than 1 to speed up the attack of the player or the mobs

- Custom.FastFishing
Set it to 1 to make fishing a lot faster!

- Custom.RespawnSpeed
Set it to a number upper than 1 to speed up the respawn of mobs (and lower than 1 to slow down).

- Custom.AutoCompleteQuestDelay
Set it to a number to enable the summon of the quest giver of the quest you just completed.
The number indicates the duration of the summoning.

- Custom.LootOnlyForPlayer
Set it to 1 to enable only certain item to be looted:
only equipment the player can be wear will be looted
only recipes for a profession the player has will be looted
only recipes the player does not already know will be looted

- Custom.LootUnique
Players can loot equipment items only once.

- Custom.MinimumRate.Drop.Item
Set it to a number to set the minimum drop rates for items of the corresponding quality